requirements.txt
pyyaml>=6.0
sqlalchemy>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0  # Für PostgreSQL
# pymysql>=1.1.0  # Optional: Für MySQL
